+++ src/javascript/components/UserInfoMain/index.js
@@ -1,0 +1,1 @@
+export { default as UserInfoMain } from './userInfoMain.js';
